pyproject.toml
[project]
name = "jenkins-job-insight"
version = "0.1.0"
description = "Jenkins job insight and analysis tool"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn",
    "python-jenkins",
    "google-genai",
    "anthropic[vertex]",
    "google-auth",
    "httpx",
    "pydantic-settings",
    "gitpython",
    "aiosqlite",
]

[project.scripts]
jenkins-job-insight = "jenkins_job_insight.main:run"

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio", "httpx"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]

[tool.hatch.build.targets.wheel]
packages = ["src/jenkins_job_insight"]
